Update local enviroment script
